pyproject.toml
[tool.poetry]
name = "nonebot-plugin-learn"
version = "0.1.0"
description = "A plugin to learn NoneBot2"
authors = ["MingxuanGame <MingxuanGame@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.5"
nonebot-adapter-onebot = "^2.1.3"
nonebot-plugin-imageutils = "^0.1.8"
nonebot-plugin-docs = "^2.0.0-beta.5"  # Exercise 2-1
nonebot-adapter-console = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"

[tool.black]
line-length = 79
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 79
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun.mirrors"
url = "https://mirrors.aliyun.com/pypi/simple/"
